Feat: Add Support for import and export of the PEQ
This commit adds both import and export of the ParametricEQ DSP Filter
as Equalizer APO compatible files.

Since Equalizer Presets in APO's file format are widely adapted, this
also allows importing the user to import the EQ generated by REW or
similar software.

For now some advanced PEQ filter types are not supported.

Preamp is also not supported and set to 0dB.