Resolve method chaining issue by splitting calls
The app didn't compile after the most recent refactor; the new line trying to use .show() on the window in tandem with set_focus wasn't chained in a way that would work. I'm not well-versed in rust, but based on the code below this, splitting out the calls seemed to make sense and compiles and functions as expected. But happy to adjust and learn more.

This was in service to trying to compile with an older glibc version (2.35.0) to prove it's possible, re: https://github.com/music-assistant/tauri-companion-app/issues/63. It seems to be working just fine compiling for the older version. Should probably reduece the dependency on later versions of glibc since they're not strictly necessary, and it would improve compatibility with older systems.